# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
...
  internal_modules_member.GroupSaveRequest:
    properties:
      id:
        type: integer
      name:
        type: string
      remark:
        type: string
      status:
        type: integer
    required:
    - name
    type: object
  internal_modules_member.LevelSaveRequest:
    properties:
      backgroundUrl:
        type: string
      discountPercent:
        type: integer
      experience:
        type: integer
      icon:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      status:
        type: integer
    required:
    - level
    - name
    type: object
  internal_modules_member.TagSaveRequest:
    properties:
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
    required:
    - name
    type: object
  internal_modules_member.UserSaveRequest:
    properties:
      areaId:
        type: integer
      avatar:
        type: string
      birthday:
        type: integer
      email:
        type: string
      groupId:
        type: integer
      id:
        type: integer
      levelId:
        type: integer
      mark:
        type: string
      mobile:
        type: string
      name:
        type: string
      nickname:
        type: string
      remark:
        type: string
      sex:
        type: integer
      status:
        type: integer
      tagIds:
        items:
          type: integer
        type: array
    required:
    - nickname
    type: object
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /member/group/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Group
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.GroupSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a member group
      tags:
      - Member Group
  /member/group/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a member level, group or tag
      tags:
      - Member Administration
  /member/group/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a member group
      tags:
      - Member Group
  /member/group/list-all-simple:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled member groups
      tags:
      - Member Group
  /member/group/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page member groups
      tags:
      - Member Group
  /member/group/simple-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled member groups
      tags:
      - Member Group
  /member/group/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Group
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.GroupSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a member group
      tags:
      - Member Group
...
  /member/level/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Level
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.LevelSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a member level
      tags:
      - Member Level
  /member/level/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a member level, group or tag
      tags:
      - Member Administration
  /member/level/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a member level
      tags:
      - Member Level
  /member/level/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List member levels
      tags:
      - Member Level
  /member/level/list-all-simple:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled member levels
      tags:
      - Member Level
  /member/level/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page member levels
      tags:
      - Member Level
  /member/level/simple-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled member levels
      tags:
      - Member Level
  /member/level/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Level
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.LevelSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a member level
      tags:
      - Member Level
...
  /member/tag/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.TagSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a member tag
      tags:
      - Member Tag
  /member/tag/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a member level, group or tag
      tags:
      - Member Administration
  /member/tag/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a member tag
      tags:
      - Member Tag
  /member/tag/list-all-simple:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled member tags
      tags:
      - Member Tag
  /member/tag/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page member tags
      tags:
      - Member Tag
  /member/tag/simple-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled member tags
      tags:
      - Member Tag
  /member/tag/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Tag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.TagSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a member tag
      tags:
      - Member Tag
  /member/user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.UserSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create an APP member
      tags:
      - Member User
  /member/user/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete an APP member
      tags:
      - Member User
  /member/user/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get an APP member
      tags:
      - Member User
  /member/user/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page APP members
      tags:
      - Member User
  /member/user/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_member.UserSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update an APP member
      tags:
      - Member User
  /member/user/update-level:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Change a member's level
      tags:
      - Member User
  /member/user/update-point:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Adjust a member's points
      tags:
      - Member User
  /member/user/update-status:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Change an APP member status
      tags:
      - Member User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"